refactor(backend): move ClaudeProcess and WebSocketSender to modules/claude (phase 3 step 5)
