+++ backend/modules/claude/claude-process.ts
@@ -1,9 +1,9 @@
 import { spawn, type ChildProcess } from 'node:child_process';
 import { EventEmitter } from 'node:events';
 import { randomUUID } from 'node:crypto';
-import { LineBuffer } from '../../modules/claude/line-buffer.util.js';
-import { parseStreamJsonLine } from '../../modules/claude/stream-json-parser.util.js';
-import { ClaudeArgsBuilder } from '../../modules/claude/claude-args-builder.js';
+import { LineBuffer } from './line-buffer.util.js';
+import { parseStreamJsonLine } from './stream-json-parser.util.js';
+import { ClaudeArgsBuilder } from './claude-args-builder.js';
 import type {
   ClaudeExecuteOptions,
   ClaudeExecutionMode,
