+++ backend/modules/claude/web-socket-sender.ts
